server/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
mysql-connector-python
pydantic
python-dotenv
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]
python-multipart
email-validator
pydantic[email]